add different referee each day test in SolverTest
