+++ src/tools/Util.java
@@ -1,4 +1,4 @@
-package Tools;
+package tools;
 
 import java.util.ArrayList;
