+++ src/tools/UtilTest.java
@@ -1,4 +1,4 @@
-package Tools;
+package tools;
 
 import static org.junit.Assert.*;
